+++ Canvas/server.js
@@ -1,68 +1,74 @@
-var express = require('express');
-var app = express();
-var cors = require('cors')
-var serv = require('http').Server(app);
-const PORT = 8080;
-var playerList = [];
-app.use(cors())
+import express from 'express';
+import http from 'http';
+import SocketIO from 'socket.io';
+import cors from 'cors';
+import Player from './src/js/Components/Player'; // this cause duplicated session
+
+/* Global */
+const app = express();
+const serv = http.Server(app);
+const PORT = process.env.PORT || 8080;
+let playerList = {};
+let SOCKET_LIST = {};
+/* Hosting */
 
 app.get('/',(req,res) => {
     res.sendFile(__dirname + '/public/index.html');
 });
-
 app.use('/dist', express.static(__dirname + '/public/dist'));
 
-serv.listen(PORT);
+/* Cors setting */
 
+app.use(cors())
 app.use(function(req, res, next) {
     res.header("Access-Control-Allow-Origin", "*");
     res.header("Access-Control-Allow-Headers", "Origin, X-Requested-With, Content-Type, Accept");
     next();
-  });
+});
 
-// var socket = require('socket.io-client')('http://localhost:8080');
-var io = require('socket.io')(serv, {log:false, origins:'*:*'});
-/* TODO: Refresh the position of players somehow for each move. 
- *
- *
-*/
-console.log(`[+] Game is running on ${PORT}`);
-
-io.sockets.on('connection', socket => {
-
-        var thisPlayer = {
-            id: socket.client.id,
-            name: "Player" + Math.floor(Math.random() * 10)
-        }
-
-    socket.emit("myPlayer", thisPlayer);
-
-        console.log(`Player: ${thisPlayer.id} joined the game`);
-
-        console.log("Number of player: " + playerList.length);
-
-    socket.on("addPlayer", player => {
-        playerList.push(player);
-        console.log(playerList);
-    })
-
-    socket.emit("getPlayer", playerList);
-
-    socket.on("playerMove", data => {
-        playerList = playerList.map(player => {
-            if(player.id == data.id) {
-                player.x = data.x;
-                player.y = data.y;
-            }
-        })
-        socket.emit("getPlayer", playerList);
-    })
-
-    socket.on("disconnect", () => {
-        playerList = playerList.filter(player=> player.id !== thisPlayer.id);
-        console.log(`Player: ${thisPlayer.id} left the game`);
-        console.log("Number of player: " + playerList.length);
-    })
-    
-
+/* SocketIO */
+var io = new SocketIO(serv, {
+    log: false, origins: '*:*'
 })
+
+io.on('connection', socket => {
+    /* Init player to activate driver.js*/
+    var thisPlayer = {
+        id: socket.client.id,
+        x: 0,
+        y: 0,
+        name: "Test",
+        npc: false
+    }
+    SOCKET_LIST[thisPlayer.id] = socket;
+    playerList[thisPlayer.id] = thisPlayer;
+    socket.emit('initPlayer', thisPlayer)
+    console.log(`Player: ${thisPlayer.id} joined the game`);
+
+    /* Player moves */
+    socket.on("move", usr => {
+        /* socketID must match, just send back socketID for dataID*/
+        console.log(usr);
+        playerList[usr.id].x = usr.x;
+        playerList[usr.id].y = usr.y;
+    })
+
+    /* Player leaves the game*/
+    socket.on("disconnect", () => {
+        delete SOCKET_LIST[thisPlayer.id];
+        delete playerList[thisPlayer.id];
+    })
+})
+
+/* Run server*/
+serv.listen(PORT, () => {
+    console.log(`Server is running on port: ` + PORT);
+});
+
+/* Update game per 0.25 second */
+setInterval(function(){
+    for(var i in SOCKET_LIST){
+        var socket = SOCKET_LIST[i];
+        socket.emit('update',playerList);
+    }
+},1000/25);
