+++ Canvas/src/js/Components/Player.js
@@ -1,8 +1,8 @@
 import Component from './Component';
 // import GameEngine from '../GameEngine'; // cause socket duplication
 export default class Player extends Component{
-    constructor(id, x, y, name, size) {
-        super(id,x,y,name, false, size);
+    constructor(id, x, y, name, mapComponent) {
+        super(id,x,y,name, false, mapComponent);
         this.alive = true;
         super.control(true);
     }
@@ -14,9 +14,4 @@ export default class Player extends Component{
     getPosition() {
         return super.getPosition();
     }
-
-    // render() {
-    //     this.object.fillStyle=GameEngine.getColor().player;
-    //     super.render();
-    // }
 }
