+++ Canvas/src/js/Global.js
@@ -1,0 +1,13 @@
+/* Store all the constant */
+export default class Global {
+    static getColor() {
+        return {
+            path: "#77B6EA",
+            block: "#37393A",
+            player: "#A6B1E1",
+            monster: "#A5243D",
+            border: "#000",
+            name: "#000"
+        }
+    }
+}
