+++ Canvas/start.js
@@ -4,4 +4,4 @@ require('babel-register')({
 })
 
 // Import the rest of our application.
-module.exports = require('./sv.js')
+module.exports = require('./server.js')
